+++ packages/webview_flutter_lwe/CHANGELOG.md
@@ -1,7 +1,8 @@
-## NEXT
+## 0.3.1
 
 * Fix new lint warnings.
 * Update minimum Flutter and Dart version to 3.13 and 3.1.
+* Supports multiple JavaScriptChannel method call.
 
 ## 0.3.0